Moving everything related to the options and settings pages in a different file and class
